backend/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
idna==3.13
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
stripe==15.1.0
typing_extensions==4.15.0
urllib3==2.6.3
psycopg2-binary==2.9.9
dj-database-url==2.1.0
cloudinary==1.38.0
django-cloudinary-storage==0.1.2
whitenoise==6.6.0
gunicorn==21.2.0
